feat : 공격판정 구현

+++ Assets/Scripts/Player/Data/PlayerStatusData.cs
@@ -9,17 +9,17 @@ public class PlayerStatusData : ScriptableObject, ISerializationCallbackReceiver
     [Header("Ä³¸¯ÅÍ ½ºÅÝ")]
     [SerializeField] float _maxHP;
     [SerializeField] float _deffense;
-    [SerializeField] float _attackPower;
+    [SerializeField] int _attackPower;
     [SerializeField] float _attackSpeed;
 
     public float maxHP { get { return _maxHP; } }
     public float deffense { get { return _deffense; } }
-    public float attackPower { get { return _attackPower; } }
+    public int attackPower { get { return _attackPower; } }
     public float attackSpeed { get { return _attackSpeed; } }
 
     [NonSerialized] public float MaxHp;
     [NonSerialized] public float Deffense;
-    [NonSerialized] public float AttackPower;
+    [NonSerialized] public int AttackPower;
     [NonSerialized] public float AttackSpeed;
 
     public void OnBeforeSerialize()
